Quality of Life Improvements
Namegen allows you to specify a filename at the prompt by typing the name of the file you want to save to.  The file gets saved to "generated/<filename>[.txt]"  Namegen now also strips all non-letter and space characters from input, so random punctuation isn't stored in generation if you feed it a bunch of sentences as input.

Namegen_Syllable also allows you to specify [s], and prompts you to spell the name you're saving, and the filename to save it to in separate prompts.  It saves both the name and syllabification of the generated word. It follows the same saving convention as Namegen.

"Cleanup" sorts words that might have arbitrary syllabification and duplicates entries.  It's currently tailored to my own setup, so it has hardcoded checks to scan folders called "generated" (location of saved names when you specify a file) and "resource" (collection of input file lists for organizational purposes).